% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxis-functions.R
\name{filter_taxis}
\alias{filter_taxis}
\title{Filtrer les trajets de taxi}
\usage{
filter_taxis(data, pickup_borough = NULL, payment = NULL)
}
\arguments{
\item{data}{Un data frame, généralement `taxis_reduit`.}

\item{pickup_borough}{Une chaîne de caractères ou `NULL`.}

\item{payment}{Une chaîne de caractères ou `NULL`.}
}
\value{
Un tibble / data frame filtré.
}
\description{
Filtre le jeu de données `taxis_reduit` selon le borough de prise en charge
et le type de paiement.
}
